🚚 [Move]: #178 graphics-state operators の index.* を graphics-state-operators.* へ rename
Refs #178
