+++ packages/core/src/content-stream/operators/graphics-state/graphics-state-operators.ts
